feat: Add training scripts, fix bugs, and enhance training infrastructure
- Add training scripts for TinyStories with configurable hyperparameters
  - Support command-line arguments and environment variables
  - Auto-calculate MAX_STEPS based on batch_size to maintain total tokens
  - Generate checkpoint directories with hyperparameter names
  - Set wandb_name from checkpoint directory name

- Fix critical bugs:
  - Fix dtype issue in data_loader.py (add .long() for token indices)
  - Fix device mismatch in multihead_attention.py (create mask on correct device)

- Enhance training.py:
  - Add --beta1 and --beta2 command-line arguments support

- Add comprehensive documentation:
  - Chinese docs for dataloader strategy, checkpointing, generation
  - Training script usage guide (README.md in scripts/)
  - Generation strategy documentation

- Add generation utilities:
  - generate.py: Command-line text generation script
  - decode.py: Core decoding functions with multiple sampling strategies

+++ cs336_basics/training/data_loader.py
@@ -81,11 +81,13 @@ def get_batch(
     # extract sequences: inputs[i] = data[start:start+context_length]
     inputs = np.array([data[i:i+context_length] for i in start_positions])
     # convert to PyTorch tensors and move to device
-    inputs = torch.from_numpy(inputs).to(device) # (batch_size, context_length)
+    # Use dtype=torch.long for token indices (required for embedding lookup)
+    inputs = torch.from_numpy(inputs).long().to(device) # (batch_size, context_length)
 
     # extract targets: targets[i] = data[start+1:start+context_length+1]
     targets = np.array([data[i+1:i+context_length+1] for i in start_positions]) # (batch_size, context_length)
     # convert to PyTorch tensors and move to device
-    targets = torch.from_numpy(targets).to(device) # (batch_size, context_length)
+    # Use dtype=torch.long for token indices (required for embedding lookup)
+    targets = torch.from_numpy(targets).long().to(device) # (batch_size, context_length)
 
     return inputs, targets
